feat(core)!: derive number representation from the unit, end to end
Every coordinate and duration cell now stores one struct - a float64
value beside an int64 numerator/denominator pair - with both sides
populated on every non-null row. Which side is authoritative is declared
once, per field, and defaults follow the unit: discrete units are
integer, quarters and beats are exact rationals, physical and graphical
units are floats. The three places that previously each decided this
independently (timeline class variables, factory inference, loader
class variables) are gone.

One builder constructs the struct under a single policy: integers store
exactly; floats mirror as the exact dyadic rational of the double when
it fits int64, and as the nearest representable ratio above that; exact
inputs are never silently degraded, and a float never fabricates a
"nice" ratio (limit_denominator is banned library-wide, enforced by a
test). One arithmetic engine replaces four hand-rolled rational-add
implementations: it computes on the authoritative side, re-mirrors the
other, raises on unit mismatch, and lets the left operand pick the
result's representation.

Exactness survives end to end: stamps carry the canonical value on the
exact lane (get_coordinate, get_unit, conversion) while get(),
subscripts and raw tables remain the documented float lane; conversion
maps built from loader columns receive exact parameters; 160 ticks at
480 PPQ converts to exactly 1/3 quarters. Values built through a scalar
constructor take the unit's representation, while values the library
derives - conversions, offsets, interpolated positions, query axes -
keep the representation their computation produced, preserving the
distinction between counted and measured quantities. Display is unified
behind one formatter that renders exact values exactly (19/2 quarters),
floats as floats, and discrete units as integers.

BREAKING CHANGE: RationalField, CoordinateParser and the public
DISCRETE_UNITS constant are removed; storage/parsing.py and
timelines/engines/coordinate_ops.py are gone. NumberType defaults are
now derived from the unit rather than per-timeline/per-loader settings.
to_int() defaults to "round" (was "truncate"). Stamp.axis carries the
canonical numeric value instead of always-float. Scalar constructors
coerce to the unit's representation, so Coordinate(1.5, quarters) now
holds Fraction(3, 2); a non-integral exact fraction in a discrete unit
raises instead of silently rounding.

+++ tests/loader/test_store.py
@@ -22,7 +22,7 @@ class TestEventDataCreation:
         store = EventData.empty(TimeUnit.ticks)
         assert len(store) == 0
         assert store.unit == TimeUnit.ticks
-        assert store.number_type == NumberType.float
+        assert store.number_type == NumberType.int
 
     def test_empty_with_number_type(self) -> None:
         """Can specify number_type for empty store."""
@@ -204,7 +204,7 @@ class TestEventDataProperties:
 
     def test_number_type_property(self, store_with_instants: EventData) -> None:
         """number_type property returns NumberType."""
-        assert store_with_instants.number_type == NumberType.float
+        assert store_with_instants.number_type == NumberType.int
 
     def test_count_property(self, store_with_instants: EventData) -> None:
         """count property returns number of events."""
